% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{explodeMatch}
\alias{explodeMatch}
\title{Explode a Match interaction into a data frame}
\usage{
explodeMatch(resp, vals.as.numeric = T, columns.as.variables = F)
}
\arguments{
\item{resp}{The response column from TAO or a matchObject from \code{\link[=makeMatch]{makeMatch()}}}

\item{vals.as.numeric}{Convert values to numeric (by removing letters and other non-numeric parts of value)}

\item{columns.as.variables}{Boolean. Set to use columns as variables instead of rows.}
}
\value{
Returns a data.frame
}
\description{
Explode a Match interaction into a data frame
}
\details{
This function is used when a Match interaction is used to ask a number of questions (in rows) with shared response options (in columns)
}
\examples{
matchresponse<-c("[agree q1; disagree q2; agree q3; disagree q4]","[agree q1; disagree q2; agree q3; agree q4]")
explodeMatch(matchresponse,vals.as.numeric=F)
}
