% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{explodeMC}
\alias{explodeMC}
\title{Explode a Multiple Choice interaction into a data frame}
\usage{
explodeMC(resp)
}
\arguments{
\item{resp}{The response column from TAO or a MCObject from \code{\link[=makeMC]{makeMC()}}}
}
\value{
Returns a data.frame
}
\description{
Explode a Multiple Choice interaction into a data frame
}
\details{
This function is used when a Multiple Choice interaction allows to choose more than one option
}
\examples{
mcresp<-c("['choice_1'; 'choice_5']","['choice_5'; 'choice_6'; 'choice_7']")
explodeMC(mcresp)
}
