% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{findRows}
\alias{findRows}
\title{find the pattern in the row}
\usage{
findRows(gantt, a = gantt$names[1], type = "used", stat = "max")
}
\arguments{
\item{gantt}{}

\item{a}{}

\item{type}{"used"/"pause"/"pattern"}

\item{stat}{max, min, sum, length}
}
\value{
used: filled spots, pause: unfilled spots between filled spots, pattern: show the pattern (returns a list)
}
\description{
find the pattern in the row
}
\examples{
response<-"{'response':'1/8 10:30 - 1/8 12:30;1/8 13:00 - 1/8 14:00;1/8 11:30 - 1/8 12:30'}"
gantt<-makeGantt(response,names=c("waitress","actor","pianist"),timespan=30,time.format="\%d/\%m \%H:\%M")
findRows(gantt,"actor")
}
