% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.method.R
\name{get.q.df}
\alias{get.q.df}
\title{Produce a data.frame from a open.pci.q object}
\usage{
get.q.df(q)
}
\arguments{
\item{q}{A open.pci.q object}
}
\value{
Returns a data.frame with statements as rows and persons as columns. Each statement has a score on each person equal to the row it is put in by the person. The upper row has the highest value, the bottom row has the value 1. If a statement has not been put in a row, it gets 0 on this person.
}
\description{
Produce a data.frame from a open.pci.q object
}
\examples{
qjson<-c("[[\"Running\"],[\"Walking\",\"Crawling\"],[\"Jumping\",\"Sprinting\",\"Jogging\"],[\"Strolling\",\"\",\"Standing\",\"Sitting\"]]",
         "[[\"Sitting\"],[\"Strolling\",\"Jogging\"],[\"Standing\",\"Sprinting\",\"Crawling\"],[\"Walking\",\"Running\",\"Jumping\"]]")
q<-get.q(qjson)
get.q.df(q)
}
