% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.method.R
\name{get.q}
\alias{get.q}
\title{Extract Q-method responses from JSON}
\usage{
get.q(qjson)
}
\arguments{
\item{qjson}{A response vector from the OpenPCI wordrank PCI}
}
\value{
Returns a list of responses, each consisting of a list ranked rows
}
\description{
Extract Q-method responses from JSON
}
\examples{
qjson<-c("[[\"Running\"],[\"Walking\",\"Crawling\"],[\"Jumping\",\"Sprinting\",\"Jogging\"],[\"Strolling\",\"\",\"Standing\",\"Sitting\"]]",
         "[[\"Sitting\"],[\"Strolling\",\"Jogging\"],[\"Standing\",\"Sprinting\",\"Crawling\"],[\"Walking\",\"Running\",\"Jumping\"]]")
q<-get.q(qjson)
q
}
