% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.method.R
\name{get.q.rows}
\alias{get.q.rows}
\title{Collect Q-method-data from respondents into rows}
\usage{
get.q.rows(q)
}
\arguments{
\item{q}{A Q-method list (from \code{\link[=get.q]{get.q()}})}
}
\value{
Returns a list with one list element for each row in the wordrank PCI, each element contains all statements in that row accross all responcents.
}
\description{
Collect Q-method-data from respondents into rows
}
\examples{
#' qjson<-c("[[\"Running\"],[\"Walking\",\"Crawling\"],[\"Jumping\",\"Sprinting\",\"Jogging\"],[\"Strolling\",\"\",\"Standing\",\"Sitting\"]]",
         "[[\"Sitting\"],[\"Strolling\",\"Jogging\"],[\"Standing\",\"Sprinting\",\"Crawling\"],[\"Walking\",\"Running\",\"Jumping\"]]")
q<-get.q(qjson)
qrows<-get.q.rows(q)
qrows
}
