% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{getDuration}
\alias{getDuration}
\title{Get duration}
\usage{
getDuration(gantt, a)
}
\arguments{
\item{gantt}{A gantt object (created by makeGantt)}

\item{a}{An element}
}
\value{
Returns duration in seconds
}
\description{
Get duration
}
\examples{
response<-matrix(c("{'response':'1/8 17:00;1/8 17:00 - 1/8 17:30;1/8 17:00 - 1/8 18:00;1/8 17:00, 1/8 18:00, 1/8 19:00, 1/8 20:00;1/8 20:30;1/8 20:00 - 1/8 20:30;1/8 19:30 - 1/8 20:30'}"))
gantt<-makeGantt(response2,names=c("waitress","actor","pianist","bartender","cleaning","ticketer","musician"),timespan=30,time.format="\%d/\%m \%H:\%M")
getDuration(gantt,"actor")
getDuration(gantt,"waitress")
getDuration(gantt,"bartender")
getDuration(gantt,"cleaning")
}
