% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{getMinTime}
\alias{getMinTime}
\alias{getMaxTime}
\title{Get the minimum/maximum start/end time among elements}
\usage{
getMinTime(gantt, a, which, strict = T)

getMaxTime(gantt, a, which, strict = T)
}
\arguments{
\item{gantt}{A gantt object (created by makeGantt)}

\item{a}{a vector of elements to compare}

\item{which}{Which time to compare ("start" or "end")}

\item{strict}{If TRUE and an element among the compared is not present, NA is returned.}
}
\value{
Returns minimum/maximum time in seconds of the earliest/latest element starting/ending
}
\description{
Get the minimum/maximum start/end time among elements
}
\examples{
response<-matrix(c("{'response':'1/8 17:00;;1/8 17:00 - 1/8 18:00;1/8 18:00, 1/8 20:00 - 1/8 21:00;1/8 20:30;1/8 20:00 - 1/8 20:30;1/8 19:30 - 1/8 20:30'}"))
gantt<-makeGantt(response,names=c("waitress","actor","pianist","bartender","cleaning","ticketer","musician"),timespan=30,time.format="\%d/\%m \%H:\%M")
getMinTime(gantt,c("pianist","bartender"),which="start")
getMaxTime(gantt,c("actor","bartender"),which="end",strict=F)
getMaxTime(gantt,"actor",which="end")
as.difftime(getMaxTime(gantt,c("pianist","bartender"),which="end")-getMinTime(gantt,c("pianist","bartender"),which="start"),units = "secs")

}
