% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getResponses}
\alias{getResponses}
\title{Get responses}
\usage{
getResponses(data, pattern = ".RESPONSE")
}
\arguments{
\item{data}{csv-formatted dataset.}

\item{pattern}{RegExp pattern of the response columns}
}
\value{
Returns a list of responses.
}
\description{
Gets responses from data produced by TAO.
}
\examples{
getResponses(data,pattern=".RESPONSE")
}
