% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getScores}
\alias{getScores}
\title{Get already scored items}
\usage{
getScores(data, pattern = ".SCORE")
}
\arguments{
\item{data}{csv-formatted dataset.}

\item{pattern}{RegExp pattern of the scoring columns.}
}
\value{
Gets scored data produced by TAO.
}
\description{
Get already scored items
}
\examples{
getScores(data,pattern=".SCORE")

}
