% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{iniResult}
\alias{iniResult}
\title{Initialize result data.frame}
\usage{
iniResult(data)
}
\arguments{
\item{data}{a data.frame or matrix with a Test.taker column.}
}
\value{
a data.frame for the results
}
\description{
Initialize result data.frame
}
\examples{
iniResult(data)
}
