% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{isBefore}
\alias{isBefore}
\alias{isFirst}
\alias{isAfter}
\alias{isLast}
\alias{isOverlap}
\alias{noOverlap}
\title{Check relations between elements: isBefore, isAfter, isOverlap, noOverlap}
\usage{
isBefore(
  gantt,
  a = gantt$names[1],
  b = gantt$names[2],
  which.a = "all",
  which.b = "all",
  orEqual = FALSE,
  strict = TRUE,
  strictAfter = TRUE
)

isFirst(
  gantt,
  a = gantt$names[1],
  which = "all",
  orEqual = FALSE,
  strict = TRUE
)

isAfter(
  gantt,
  a = gantt$names[1],
  b = gantt$names[2],
  which.a = "all",
  which.b = "all",
  orEqual = FALSE,
  strict = TRUE
)

isLast(
  gantt,
  a = gantt$names[1],
  which = "all",
  orEqual = FALSE,
  strict = TRUE
)

isOverlap(
  gantt,
  a = gantt$names[1],
  b = gantt$names[2],
  which.a = "any",
  which.b = "any"
)

noOverlap(gantt, a = gantt$names[1], b = gantt$names[1])
}
\arguments{
\item{gantt}{gantt object (from makeGantt).}

\item{a}{element in the gantt chart (character vector).}

\item{b}{character or vector of characters}

\item{which.a}{"all": all a should be after, last: Only last needs to be after (ends later), or use a number to use in comparison}

\item{which.b}{"all": all b should be before, first: Only first b needs to be after (starts before), or use a number to use in comparison}

\item{orEqual}{if TRUE, a and b can be at the same time (not just before/after).}

\item{strict}{if FALSE, an element a is considered before/after even though b is not present.}

\item{stict}{when TRUE, isAfter returns 0 when b is not set, otherwise a is after b, when b is not set}

\item{which.a/which.b}{any: one or more a/b elements overlap, all: all a/b elements need to overlap, first/first: First a/b needs to overlap, last/last: Last a/b needs to overlap, or use number (or sequence)}
}
\value{
When stict is TRUE, isBefore returns 0 when b is not set, otherwise a is before b, when b is not set
}
\description{
Check relations between elements: isBefore, isAfter, isOverlap, noOverlap

Title
}
\examples{
response<-"{'response':'1/8 10:30 - 1/8 12:30;1/8 13:00 - 1/8 14:00;1/8 11:30 - 1/8 12:30'}"
gantt<-makeGantt(response,names=c("waitress","actor","pianist"),timespan=30,time.format="\%d/\%m \%H:\%M")
isBefore(gantt,"actor",c("waitress","pianist"),which.a = "first",which.b = "first",orEqual = TRUE,strict = FALSE)
isAfter(gantt,"actor",c("waitress"),which.a = "last",which.b = "all")
isLast(gantt,"actor")
isFirst(gantt,"actor")
isOverlap(gantt,"actor","waitress")
}
