% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{isOrderBefore}
\alias{isOrderBefore}
\alias{isOrderAfter}
\alias{isOrderSet}
\title{Test ordering of elements}
\usage{
isOrderBefore(orderObject, a, b, strict = T, count = F)

isOrderAfter(orderObject, a, b, strict = T)

isOrderSet(orderObject, a)
}
\arguments{
\item{orderObject}{An orderObject}

\item{a}{an element}

\item{b}{an element}

\item{strict}{All members of b needs to be present}

\item{count}{How many in b is a before?}
}
\value{
Returns scores
}
\description{
Test ordering of elements
}
\examples{
orderresp<-c("<'audition'; 'actor'; 'prof reading'; 'costumes'>", "<'manuscript'; 'prof reading'; 'costumes'; 'actor'>")
orderObject<-makeOrder(orderresp)
isOrderSet(orderObject,"audition")+isOrderBefore(orderObject,"audition",c("costumes", "prof reading"),count = T)
}
