% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{isSet}
\alias{isSet}
\title{Has this element been given a time slot}
\usage{
isSet(gantt, a = 1)
}
\arguments{
\item{gantt}{A gantt object (created by makeGantt)}

\item{a}{The name or row number of the element}
}
\value{
Returns a vector of booleans
}
\description{
Has this element been given a time slot
}
\examples{
response<-"{'response':'1/8 10:30 - 1/8 12:30;1/8 13:00 - 1/8 14:00;1/8 11:30 - 1/8 12:30'}"
gantt<-makeGantt(response,names=c("waitress","actor","pianist"),timespan=30,time.format="\%d/\%m \%H:\%M")
isSet(gantt,a="actor")
}
