% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{makeGantt}
\alias{makeGantt}
\title{Make Gantt Object}
\usage{
makeGantt(gantt, names, timespan = 30, time.format = "\%d/\%m \%H:\%M")
}
\arguments{
\item{gantt}{The Gantt data from TAO}

\item{names}{Names of the rows in the Gantt chart (don't have to be the same as in the PCI)}

\item{timespan}{Duration of a time slot}

\item{time.format}{The POSSIX time format that the data was formatted (see https://www.stat.berkeley.edu/~s133/dates.html). Use \\\%d/0 when the Gantt-output was made using 0 for month}
}
\value{
Returns a Gantt object
}
\description{
Make Gantt Object
}
\examples{
response<-"{'response':'1/8 10:30 - 1/8 12:30;1/8 13:00 - 1/8 14:00;1/8 11:30 - 1/8 12:30'}"
makeGantt(response,names=c("waitress","actor","pianist"),timespan=30,time.format="\%d/\%m \%H:\%M")
}
