% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{makeMC}
\alias{makeMC}
\title{Make a multiple choise object}
\usage{
makeMC(resp)
}
\arguments{
\item{resp}{A response vector from TAO}
}
\value{
Returns a multiple choice object
}
\description{
Make a multiple choise object
}
\examples{
mcresp<-c("['choice_1'; 'choice_5']","['choice_5'; 'choice_6']")
mc<-makeMC(mcresp)
}
