% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{makeSC}
\alias{makeSC}
\title{Make a single choise object}
\usage{
makeSC(resp)
}
\arguments{
\item{resp}{A response vector from TAO}
}
\value{
Returns a multiple choice object
}
\description{
Make a single choise object
}
\examples{
scresp<-c("choice_1","choice_5")
sc<-makeSC(scresp)
}
