% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textGapMatch.R
\name{makeTextGapMatch}
\alias{makeTextGapMatch}
\title{Make a text gap match object}
\usage{
makeTextGapMatch(resp)
}
\arguments{
\item{resp}{A response vector from TAO}
}
\value{
Returns a text gap match object
}
\description{
Make a text gap match object
}
\examples{
mcresp<-c("['choice_1'; 'choice_5']","['choice_5'; 'choice_6']")
mc<-makeMC(mcresp)
}
