% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{numSlots}
\alias{numSlots}
\title{numSlots}
\usage{
numSlots(gantt, a = gantt$names[1])
}
\arguments{
\item{gantt}{a gantt object (created by makeGantt)}

\item{a}{an element}
}
\value{
Returns number of time slots - also counting non-connected slots
}
\description{
numSlots
}
\examples{
response<-"{'response':'1/8 10:30 - 1/8 12:30;1/8 13:00 - 1/8 14:00;1/8 11:30 - 1/8 12:30'}"
gantt<-makeGantt(response,names=c("waitress","actor","pianist"),timespan=30,time.format="\%d/\%m \%H:\%M")
numSlots(gantt,"actor")
}
