% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textGapMatch.R
\name{posInGap}
\alias{posInGap}
\title{The position in gap match}
\usage{
posInGap(textGapMatchObject, text, gap)
}
\arguments{
\item{textGapMatchObject}{a textGapMatchObject}

\item{text}{The text to look for}

\item{gap}{The gap to look in}
}
\value{
Returns the scores
}
\description{
The position in gap match
}
\examples{
resp<-"{\"dropzone_1\":[\"1.Decide size of building\",\"2.Paint the walls\",\"4.Get a quote\"],\"dropzone_2\":[\"1.Move furniture\",\"2.Hire workers\"],\"dropzone_3\":[\"1.Test the construction\",\"2.Move in\"]}"
gm<-makeTextGapMatch(resp)
posInGap(gm,"Paint the walls",1)
}
