% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.method.R
\name{q.distribution}
\alias{q.distribution}
\title{Tabularize Q-method-data}
\usage{
q.distribution(qrows, exclude = NULL)
}
\arguments{
\item{qrows}{A list of rows from \code{\link[=get.q.rows]{get.q.rows()}}}

\item{exclude}{A vector of values to exclude from the tabulation}
}
\value{
Returns a data.frame with three rows: Activity (the statements from wordrank), Frequency of this statement in this row, and the Row for which the Activity has this Frequency. Rows are numbered from the buttom (1) to the top (number of rows)
}
\description{
Tabularize Q-method-data
}
\examples{
qjson<-c("[[\"Running\"],[\"Walking\",\"Crawling\"],[\"Jumping\",\"Sprinting\",\"Jogging\"],[\"Strolling\",\"\",\"Standing\",\"Sitting\"]]",
         "[[\"Sitting\"],[\"Strolling\",\"Jogging\"],[\"Standing\",\"Sprinting\",\"Crawling\"],[\"Walking\",\"Running\",\"Jumping\"]]")
q<-get.q(qjson)
qrows<-get.q.rows(q)
qdist<-q.distribution(qrows)
qdist
}
\seealso{
\code{\link[=q.plot]{q.plot()}}
}
