% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.method.R
\name{q.plot}
\alias{q.plot}
\title{Plot distribution of Q-method ordering}
\usage{
q.plot(qdist, title = NULL, subtitle = NULL, caption = NULL)
}
\arguments{
\item{qdist}{The distributions of Q-method ordering from \code{\link[=qdist]{qdist()}}}

\item{title}{The title of the plot}

\item{subtitle}{The subtitle of the plot}

\item{caption}{The caption of the plot}
}
\value{
Returns the plot (a ggplot which for example can be saved with \code{\link[=ggsave]{ggsave()}})
}
\description{
Plot distribution of Q-method ordering
}
\examples{
qjson<-c("[[\"Running\"],[\"Walking\",\"Crawling\"],[\"Jumping\",\"Sprinting\",\"Jogging\"],[\"Strolling\",\"\",\"Standing\",\"Sitting\"]]",
         "[[\"Sitting\"],[\"Strolling\",\"Jogging\"],[\"Standing\",\"Sprinting\",\"Crawling\"],[\"Walking\",\"Running\",\"Jumping\"]]")
q<-get.q(qjson)
qrows<-get.q.rows(q)
qdist<-q.distribution(qrows)
qplot(qdist)
}
