% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.method.R
\name{q.show.distribution}
\alias{q.show.distribution}
\title{Show distribution of statements in factors}
\usage{
q.show.distribution(QmethodRes, marg = 0.03, linewidth = 15, show.plot = T)
}
\arguments{
\item{QmethodRes}{A result object from \code{\link[=qmethod]{qmethod()}} or a data.frame with statements as rownames and factors in columns with position of statements.}

\item{marg}{Margin of rectangles (relative to size of retangle)}

\item{linewidth}{Linewidth of texts in number of letters (longer lines are split at nearest space)}

\item{show.plot}{If TRUE, plot is output to active dev.}
}
\value{
Returns a list of the produced plots.
}
\description{
Show distribution of statements in factors
}
\examples{
qjson<-c("[[\"Running\"],[\"Walking\",\"Crawling\"],[\"Jumping\",\"Sprinting\",\"Jogging\"],[\"Strolling\",\"\",\"Standing\",\"Sitting\"]]",
         "[[\"Sitting\"],[\"Strolling\",\"Jogging\"],[\"Standing\",\"Sprinting\",\"Crawling\"],[\"Walking\",\"Running\",\"Jumping\"]]")
q<-get.q(qjson)
result<-q.analysis(q)
q.show.distribution(result)
}
