% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{removeDuplicatedIds}
\alias{removeDuplicatedIds}
\title{Remove duplicated ids}
\usage{
removeDuplicatedIds(result, idcol = colnames(result)[1])
}
\arguments{
\item{result}{data.frame of scoring results.}

\item{idcol}{the name of the column that contains person id's. If none given, the first column of \code{result} is used.}
}
\value{

}
\description{
Remove duplicated ids
}
\details{
Duplicated persons are combined into one.

Scores on all variables are set to the max score given to this person on this variable. If there are only NA's on a variable, NA is retained. Non-numeric variables are given the value of the first occurrence of the person.
}
\examples{
result<-data.frame(id=c("b","b","a","a","c","d","a"),X1_result=c(1,NA,2,2,1,3,NA),X2_party=c(2,1,3,1,2,3,NA),X20_dont=c(NA,2,1,2,1,3,NA),sex=c("Male","Female","Male","Female","Female","Male","Female"))
removeDuplicatedIds(result)
}
