% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{scoreSC}
\alias{scoreSC}
\title{Score single or multiple choice objects}
\usage{
scoreSC(mc, correct = c())
}
\arguments{
\item{mc}{a multiple or single choice object}

\item{correct}{the correct response(s)}
}
\value{
returns a vector of scores
}
\description{
Score single or multiple choice objects
}
\examples{

mcresp<-c("['choice_1'; 'choice_5']","['choice_5'; 'choice_6']")
mc<-makeMC(mcresp)
scoreMC(mc,correct=c("choice_1","choice_5"))
}
