% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textGapMatch.R
\name{textInGap}
\alias{textInGap}
\title{Test for text in gap match}
\usage{
textInGap(textGapMatchObject, text, gap, no.where.else = TRUE)
}
\arguments{
\item{textGapMatchObject}{a textGapMatchObject}

\item{text}{The text to look for}

\item{gap}{The gap(s) to look in}

\item{no.where.else}{Only accept it if it is only here}
}
\value{
Returns the scores
}
\description{
Test for text in gap match
}
\examples{
resp<-"{\"dropzone_1\":[\"1.Decide size of building\",\"2.Paint the walls\",\"4.Get a quote\"],\"dropzone_2\":[\"1.Move furniture\",\"2.Hire workers\"],\"dropzone_3\":[\"1.Test the construction\",\"2.Move in\",\"3.Paint the walls\"]}"
gm<-makeTextGapMatch(resp)
textInGap(gm,"Paint the walls","dropzone_1")
textInGap(gm,"Paint the walls","dropzone_1",no.where.else=F)
}
