% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{writeResult}
\alias{writeResult}
\alias{writeCoded}
\title{Write results}
\usage{
writeResult(result, filename, resultsdir, exclude = c())

writeCoded(coded, filename, resultsdir)
}
\arguments{
\item{result}{A data.frame with results}

\item{filename}{Where to write the results}

\item{resultsdir}{The path to the directory of the file}

\item{exclude}{A vector of column names or numbers which are not numeric, and therefore should be excluded from the cleaning.}
}
\value{
Returns the cleaned results.
}
\description{
Write results
}
\examples{
writeResult(result,filename,resultsdir)
}
