% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{getDuration}
\alias{getDuration}
\title{Get duration}
\usage{
getDuration(gantt, a)
}
\arguments{
\item{gantt}{A gantt object (created by makeGantt)}

\item{a}{An element}
}
\value{

}
\description{
Get duration
}
\examples{
response<-matrix(c("{'response':'1/8 10:30 - 1/8 12:30;1/8 13:00 - 1/8 14:00;1/8 11:30 - 1/8 12:30'}","{'response':'1/8 12:00 - 1/8 12:30;1/8 11:00 - 1/8 12:00;1/8 10:30 - 1/8 11:00'}"))
gantt<-makeGantt(response,names=c("waitress","actor","pianist"),timespan=30,time.format="\%d/\%m \%H:\%M")
getDuration(gantt,"actor")
}
