% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{getMinTime}
\alias{getMinTime}
\alias{getMaxTime}
\title{Get the element with minimum/maximum start/end}
\usage{
getMinTime(gantt, a, which)

getMaxTime(gantt, a, which)
}
\arguments{
\item{gantt}{A gantt object (created by makeGantt)}

\item{a}{a vector of elements to compare}

\item{which}{Which time to compare ("start" or "end")}
}
\value{
Returns minimum/maximum time of the earliest/latest element starting/ending in seconds
}
\description{
Get the element with minimum/maximum start/end
}
\details{
If an element among the compared is not present, NA is returned.
}
\examples{
response<-matrix(c("{'response':'1/8 17:00;;1/8 17:00 - 1/8 18:00;1/8 18:00, 1/8 20:00 - 1/8 21:00;1/8 20:30;1/8 20:00 - 1/8 20:30;1/8 19:30 - 1/8 20:30'}"))
gantt<-makeGantt(response,names=c("waitress","actor","pianist","bartender","cleaning","ticketer","musician"),timespan=30,time.format="\%d/\%m \%H:\%M")
getMinTime(gantt,c("pianist","bartender"),which="start")
getMaxTime(gantt,"bartender",which="end")
getMaxTime(gantt,"actor",which="end")
}
