% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{getMinTime}
\alias{getMinTime}
\alias{getMaxTime}
\title{Get the element taking up the fewest/most time slots}
\usage{
getMinTime(gantt, a, which)

getMaxTime(gantt, a, which)
}
\arguments{
\item{gantt}{A gantt object (created by makeGantt)}

\item{a}{a vector of elements to compare}

\item{which}{start/end}
}
\value{
Returns a vector of minimum/maximum time
}
\description{
Get the element taking up the fewest/most time slots
}
\examples{
response<-matrix(c("{'response':'1/8 10:30 - 1/8 12:30;1/8 11:00 - 1/8 12:00;1/8 11:30 - 1/8 12:00'}","{'response':'1/8 12:00 - 1/8 12:30;1/8 11:00 - 1/8 12:00;1/8 10:30 - 1/8 11:00'}"))
gantt<-makeGantt(response,names=c("waitress","actor","pianist"),timespan=30,time.format="\%d/\%m \%H:\%M")
getMinTime(gantt,"actor",which="end")
getMaxTime(gantt,"actor",which="end")
}
