% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textGapMatch.R
\name{makeTextGapMatch}
\alias{makeTextGapMatch}
\title{Make a text gap match object}
\usage{
makeTextGapMatch(resp)
}
\arguments{
\item{resp}{A response vector from TAO}
}
\value{
Returns a text gap match object
}
\description{
Make a text gap match object
}
\examples{
matchresp<-c('{"room":["Chair"],"kitchen":["Pot"],"hallway":["Staircase"]}','{"room":["Sofa","table"],"kitchen":["Books"],"hallway":["Lion"]}')
matchObj<-makeTextGapMatch(matchresp)
}
