% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeScores.R
\name{mergeScores}
\alias{mergeScores}
\title{Merge scores from pre-coded items}
\usage{
mergeScores(result, resp, column, prefix = "", test.taker = NULL)
}
\arguments{
\item{result}{result of scoring script. A matrix.}

\item{resp}{Response data.frame from TAO or from pre-coded items.}

\item{column}{String. The beginning of the name of the columns to merge with the results. If more columns start with the provided string, all of them are included.}

\item{prefix}{Text to put before the variable name}

\item{test.taker}{vector of test taker-ids for the response data.frame. If left empty, first column in resp is used as test.taker.}
}
\value{
returns a result matrix
}
\description{
Merge scores from pre-coded items
}
\examples{
# Make sure not to import data from item X20
result<-data.frame(id=c("a","b","a","b","c","d","a"))
resp<-data.frame(id=c("b","b","a","a","c","d","a"),X1=c(1,NA,2,2,1,3,NA))

result<-mergeScores(result=result,resp=resp,column="X1",prefix="X1_")
}
