% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room.R
\name{mergeScores}
\alias{mergeScores}
\title{mergeScores}
\usage{
mergeScores(result, resp, column, prefix = "")
}
\arguments{
\item{result}{result of scoring script. A matrix.}

\item{resp}{Response data.frame from TAO or from pre-coded items.}

\item{column}{String. The beginning of the name of the columns to merge with the results. If more columns start with the provided string, all of them are included.}

\item{prefix}{Text to put before the variable name}
}
\value{
returns a result matrix
}
\description{
mergeScores
}
\examples{
result<-mergeScores(result,prescoredresults,"X2..Responses.RESPONSE","TM2_")
}
