% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeScores.R
\name{mergeScores}
\alias{mergeScores}
\title{Merge scores from pre-coded items}
\usage{
mergeScores(result, resp, column, prefix = "", test.taker = NULL)
}
\arguments{
\item{result}{result of scoring script. A matrix.}

\item{resp}{Response data.frame from TAO or from pre-coded items.}

\item{column}{String. The beginning of the name of the columns to merge with the results. If more columns start with the provided string, all of them are included.}

\item{prefix}{Text to put before the variable name}

\item{test.taker}{vector of test taker-ids for the response data.frame. If left empty, first column in resp is used as test.taker.}
}
\value{
returns a result matrix
}
\description{
Merge scores from pre-coded items
}
\examples{
# Make sure not to import data from item X20
result<-mergeScores(result,prescoredresults,"X2[^0-9]","TM2_")
}
