% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textGapMatch.R
\name{posInGap}
\alias{posInGap}
\title{The position in gap match}
\usage{
posInGap(textGapMatchObject, text, gap)
}
\arguments{
\item{textGapMatchObject}{a textGapMatchObject}

\item{text}{The text to look for}

\item{gap}{The gap to look in}
}
\value{
Returns the scores
}
\description{
The position in gap match
}
\examples{
To come
}
