% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{scoreMatch}
\alias{scoreMatch}
\title{Score match responses}
\usage{
scoreMatch(resp, identifier, correct = c(), columns.as.variables = F)
}
\arguments{
\item{identifier}{The variable that has been given (a) corresponding value(s)}

\item{correct}{The correct values}

\item{columns.as.variables}{Boolean. Set to use columns as variables instead of rows.}

\item{matchObject}{A matchObject}
}
\value{
A vector of scores
}
\description{
Score match responses
}
\examples{
matchresponse<-c("[mail1 annoyed; mail4 annoyed; mail2 angry]","[mail1 confused]")
matchObject<-makeMatch(matchresponse)
scoreMatch(matchObject,"mail1",c("happy","annoyed"))
}
