% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textGapMatch.R
\name{textInGap}
\alias{textInGap}
\title{Test for text in gap match}
\usage{
textInGap(textGapMatchObject, text, gap, no.where.else = TRUE)
}
\arguments{
\item{textGapMatchObject}{a textGapMatchObject}

\item{text}{The text to look for}

\item{gap}{The gap to look in}

\item{no.where.else}{Only accept it if it is only here}
}
\value{
Returns the scores
}
\description{
Test for text in gap match
}
\examples{
To come
}
