% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{writeResult}
\alias{writeResult}
\title{Write results}
\usage{
writeResult(result, filename, resultsdir)
}
\arguments{
\item{result}{A data.frame with results}

\item{filename}{Where to write the results}

\item{resultsdir}{The path to the directory of the file}
}
\value{
Returns the cleaned results.
}
\description{
Write results
}
\examples{
writeResult(result,filename,resultsdir)
}
